+++ Pong/src/pong/Jogador/IJogador.java
@@ -6,7 +6,7 @@ import pong.Raquete;
 public interface IJogador {
 
     // calcula a velocidade que a raquete deve se mover
-    public double calculaVelocidade(Raquete minha, Raquete oponente, Bola bola);
+    public int calculaVelocidade(Raquete minha, Raquete oponente, Bola bola);
 
     // método chamado quando alguém faz uma pontuação
     // ponto = 1 - jogador fez o ponto. ponto = -1 - oponente fez o ponto
